Updating gorm requests, api separation started.
