+++ routes/router.go
@@ -14,22 +14,21 @@ func InitRouter() *gin.Engine {
 
 	users := router.Group("/users")
 	{
-		user := new(controllers.User)
-		users.POST("/list", user.ListUsers)
+		user := new(controllers.WorkerController)
 		users.POST("/login", user.Login)
 		users.POST("/register", user.Register)
 		users.PATCH("/edit/:username", user.Update)
 		users.DELETE("/remove/:username", user.Destroy)
 		users.GET("/get/:username", user.Get)
 	}
-	game := router.Group("/inventory")
-	{
-		gameState := new(controllers.GameState)
-		game.GET("/available", gameState.GetAvailableInventory) // available items for gametype x
-		game.POST("/transfer", gameState.TransferInventory)     // from available inventory to slot of user y
-		game.PATCH("/layout/edit", gameState.UpdateUserLayout)  // change layout of user y
-		game.POST("/update", gameState.UpdateInventory)         // edit available items in inventory for gametype x
-	}
+	//game := router.Group("/inventory")
+	//{
+	//	//gameState := new(controllers.GameState)
+	//	//game.GET("/available", gameState.GetAvailableInventory) // available items for gametype x
+	//	//game.POST("/transfer", gameState.TransferInventory)     // from available inventory to slot of user y
+	//	//game.PATCH("/layout/edit", gameState.UpdateUserLayout)  // change layout of user y
+	//	//game.POST("/update", gameState.UpdateInventory)         // edit available items in inventory for gametype x
+	//}
 
 	ping := new(controllers.Ping)
 	router.GET("/ping", ping.Ping)
